Provide `std::unreachable` if not present
There were a few places in the code where the intention was to use
`std::unreachable()`. Those places were commented out due to some older
compiler not supporting this feature. This commit adds code to check the
feature-test macro `__cpp_lib_unreachable` and, if absent, it provides
an implementation of `std::unreachable()` compatible with C++23.